requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
aiohttp==3.10.8
cachetools==5.5.0
python-dotenv==1.0.1
loguru==0.7.2